added blobconverter CLI & cleanup
