+++ depthai_demo.py
@@ -1,415 +1,387 @@
 #!/usr/bin/env python3
-
+import argparse
 import json
-import platform
-import os
-from time import time, monotonic
+import time
+from pathlib import Path
+from types import SimpleNamespace
+
 import cv2
+import depthai as dai
 import numpy as np
-import depthai
-from depthai_helpers.version_check import check_depthai_version
-from depthai_helpers.object_tracker_handler import show_tracklets
-from depthai_helpers.config_manager import DepthConfigManager
-from depthai_helpers.arg_manager import CliArgs
 
-print('Using depthai module from: ', depthai.__file__)
-print('Depthai version installed: ', depthai.__version__)
-check_depthai_version()
-is_rpi = platform.machine().startswith('arm') or platform.machine().startswith('aarch64')
-global args, cnn_model2
+from depthai_helpers.arg_manager import parse_args
+from depthai_helpers.config_manager import BlobManager, ConfigManager
+from depthai_helpers.utils import frame_norm, to_planar, to_tensor_result
+
+conf = ConfigManager(parse_args())
+
+in_w, in_h = conf.getInputSize()
+rgb_res = conf.getRgbResolution()
+median = conf.getMedianFilter()
 
 
-class DepthAI:
-    global is_rpi
-    process_watchdog_timeout = 10  # seconds
-    nnet_packets = None
-    data_packets = None
-    runThread = True
+class NNetManager:
+    source_choices = ("rgb", "left", "right", "host")
+    config = None
+    nn_family = None
+    confidence = None
+    metadata = None
+    output_format = None
+    device = None
+    input = None
+    output = None
+    sbb = False
 
-    def reset_process_wd(self):
-        global wd_cutoff
-        wd_cutoff = monotonic() + self.process_watchdog_timeout
-        return
+    def __init__(self, source, use_depth, use_hq, model_dir=None, model_name=None):
+        if source not in self.source_choices:
+            raise RuntimeError(f"Source {source} is invalid, available {self.source_choices}")
 
-    def on_trackbar_change(self, value):
-        self.device.send_disparity_confidence_threshold(value)
-        return
+        self.model_name = model_name
+        self.model_dir = model_dir
+        self.source = source
+        self.use_depth = use_depth
+        self.use_hq = use_hq
+        self.output_name = f"{self.model_name}_out"
+        self.input_name = f"{self.model_name}_in"
+        self.blob_path = BlobManager(model_dir=self.model_dir, model_name=self.model_name).compile(conf.args.shaves)
 
-    def stopLoop(self):
-        self.runThread = False
+        if model_dir is not None:
+            config_path = self.model_dir / Path(self.model_name).with_suffix(f".json")
+            if config_path.exists():
+                with config_path.open() as f:
+                    self.config = json.load(f)
+                    nn_config = self.config.get("NN_config", {})
+                    self.labels = self.config.get("mappings", {}).get("labels", None)
+                    self.nn_family = nn_config.get("NN_family", None)
+                    self.output_format = nn_config.get("output_format", None)
+                    self.metadata = nn_config.get("NN_specific_metadata", {})
 
-    def startLoop(self):
-        cliArgs = CliArgs()
-        args = vars(cliArgs.parse_args())
+                    self.confidence = self.metadata.get("confidence_threshold", nn_config.get("confidence_threshold", None))
 
-        configMan = DepthConfigManager(args)
-        if is_rpi and args['pointcloud']:
-            raise NotImplementedError("Point cloud visualization is currently not supported on RPI")
-        # these are largely for debug and dev.
-        cmd_file, debug_mode = configMan.getCommandFile()
-        usb2_mode = configMan.getUsb2Mode()
+                    # Disaply depth roi bounding boxes
+                    self.sbb = conf.args.spatial_bounding_box and self.use_depth and self.nn_family in ("YOLO", "mobilenet")
 
-        # decode_nn and show_nn are functions that are dependent on the neural network that's being run.
-        decode_nn = configMan.decode_nn
-        show_nn = configMan.show_nn
+    def addDevice(self, device):
+        self.device = device
+        self.input = device.getInputQueue(self.input_name, maxSize=1, blocking=False) if self.source == "host" else None
+        self.output = device.getOutputQueue(self.output_name, maxSize=1, blocking=False)
 
-        # Labels for the current neural network. They are parsed from the blob config file.
-        labels = configMan.labels
-        NN_json = configMan.NN_config
+        if self.sbb:
+            self.sbb_out = device.getOutputQueue("sbb", maxSize=1, blocking=False)
+            self.depth_out = device.getOutputQueue("depth", maxSize=1, blocking=False)
 
-        # This json file is sent to DepthAI. It communicates what options you'd like to enable and what model you'd like to run.
-        config = configMan.jsonConfig
-
-        # Create a list of enabled streams ()
-        stream_names = [stream if isinstance(stream, str) else stream['name'] for stream in configMan.stream_list]
-
-        enable_object_tracker = 'object_tracker' in stream_names
-
-        # grab video file, if option exists
-        video_file = configMan.video_file
-
-        self.device = None
-        if debug_mode:
-            print('Cmd file: ', cmd_file, ' args["device_id"]: ', args['device_id'])
-            self.device = depthai.Device(cmd_file, args['device_id'])
+    def create_nn_pipeline(self, p, nodes):
+        if self.nn_family == "mobilenet":
+            nn = p.createMobileNetSpatialDetectionNetwork() if self.use_depth else p.createMobileNetDetectionNetwork()
+            nn.setConfidenceThreshold(self.confidence)
+        elif self.nn_family == "YOLO":
+            nn = p.createYoloSpatialDetectionNetwork() if self.use_depth else p.createYoloDetectionNetwork()
+            nn.setConfidenceThreshold(self.confidence)
+            nn.setNumClasses(self.metadata["classes"])
+            nn.setCoordinateSize(self.metadata["coordinates"])
+            nn.setAnchors(self.metadata["anchors"])
+            nn.setAnchorMasks(self.metadata["anchor_masks"])
+            nn.setIouThreshold(self.metadata["iou_threshold"])
         else:
-            self.device = depthai.Device(args['device_id'], usb2_mode)
+            # TODO use createSpatialLocationCalculator
+            nn = p.createNeuralNetwork()
 
-        print(stream_names)
-        print('Available streams: ' + str(self.device.get_available_streams()))
+        nn.setBlobPath(str(self.blob_path))
+        nn.setNumInferenceThreads(2)
+        nn.input.setBlocking(False)
+        nn.input.setQueueSize(2)
+        xout = p.createXLinkOut()
+        xout.setStreamName(self.output_name)
+        nn.out.link(xout.input)
+        setattr(nodes, self.model_name, nn)
+        setattr(nodes, self.output_name, xout)
+        if self.source == "rgb":
+            nodes.cam_rgb.preview.link(nn.input)
+        elif self.source == "host":
+            xin = p.createXLinkIn()
+            xin.setStreamName(self.input_name)
+            xin.out.link(nn.input)
+            setattr(nodes, self.input_name, xout)
+        elif self.source == "right": # Use spatial information
+            # Set XLinkOut sources
+            if conf.args.sync:
+                nn.passthrough.link(nodes.xout_right.input)
+            if self.sbb:
+                if conf.args.sync:
+                    nn.passthroughDepth.link(nodes.xout_depth.input)
+                # If we want to display spatial bounding boxes, create XLinkOut node SBBs:
+                xout_sbb = self.p.createXLinkOut()
+                xout_sbb.setStreamName("sbb")
+                nn.boundingBoxMapping.link(xout_sbb.input)
 
-        # create the pipeline, here is the first connection with the device
-        p = self.device.create_pipeline(config=config)
+            # NN inputs
+            nodes.manip.out.link(nn.input)
+            if conf.useDepth and self.nn_family in ("mobilenet", "YOLO"):
+                nodes.stereo.depth.link(nn.inputDepth)
+                # Spatial configs
+                nn.setBoundingBoxScaleFactor(conf.args.sbb_scale_factor)
+                nn.setDepthLowerThreshold(100)
+                nn.setDepthUpperThreshold(3000)
 
-        if p is None:
-            print('Pipeline is not created.')
-            exit(3)
+    def get_label_text(self, label):
+        if self.config is None or self.labels is None:
+            return label
+        elif int(label) < len(self.labels):
+            return self.labels[int(label)]
+        else:
+            print(f"Label of ouf bounds (label_index: {label}, available_labels: {len(self.labels)}")
+            return label
 
-        nn2depth = self.device.get_nn_to_depth_bbox_mapping()
 
-        t_start = time()
-        frame_count = {}
-        frame_count_prev = {}
-        nnet_prev = {}
-        nnet_prev["entries_prev"] = {}
-        nnet_prev["nnet_source"] = {}
-        frame_count['nn'] = {}
-        frame_count_prev['nn'] = {}
+class FPSHandler:
+    def __init__(self, cap=None):
+        self.timestamp = time.time()
+        self.start = time.time()
+        self.framerate = cap.get(cv2.CAP_PROP_FPS) if cap is not None else None
 
-        NN_cams = {'rgb', 'left', 'right'}
+        self.frame_cnt = 0
+        self.ticks = {}
+        self.ticks_cnt = {}
 
-        for cam in NN_cams:
-            nnet_prev["entries_prev"][cam] = None
-            nnet_prev["nnet_source"][cam] = None
-            frame_count['nn'][cam] = 0
-            frame_count_prev['nn'][cam] = 0
+    def next_iter(self):
+        if not conf.useCamera:
+            frame_delay = 1.0 / self.framerate
+            delay = (self.timestamp + frame_delay) - time.time()
+            if delay > 0:
+                time.sleep(delay)
+        self.timestamp = time.time()
+        self.frame_cnt += 1
 
-        stream_windows = []
-        for s in stream_names:
-            if s == 'previewout':
-                for cam in NN_cams:
-                    stream_windows.append(s + '-' + cam)
-            else:
-                stream_windows.append(s)
+    def tick(self, name):
+        if name in self.ticks:
+            self.ticks_cnt[name] += 1
+        else:
+            self.ticks[name] = time.time()
+            self.ticks_cnt[name] = 0
 
-        for w in stream_windows:
-            frame_count[w] = 0
-            frame_count_prev[w] = 0
+    def tick_fps(self, name):
+        if name in self.ticks:
+            time_diff = time.time() - self.ticks[name]
+            return self.ticks_cnt[name] / time_diff if time_diff != 0 else 0
+        else:
+            return 0
 
-        tracklets = None
+    def fps(self):
+        return self.frame_cnt / (self.timestamp - self.start)
 
-        self.reset_process_wd()
 
-        time_start = time()
+class PipelineManager:
+    def __init__(self, nn_manager):
+        self.p = dai.Pipeline()
+        self.nodes = SimpleNamespace()
+        self.nn_manager = nn_manager
 
-        def print_packet_info_header():
-            print('[hostTimestamp streamName] devTstamp seq camSrc width height Bpp')
+    def create_color_cam(self, use_hq):
+        # Define a source - color camera
+        self.nodes.cam_rgb = self.p.createColorCamera()
+        self.nodes.cam_rgb.setPreviewSize(in_w, in_h)
+        self.nodes.cam_rgb.setInterleaved(False)
+        self.nodes.cam_rgb.setResolution(rgb_res)
+        self.nodes.cam_rgb.setFps(conf.args.rgb_fps)
+        xout_rgb = self.p.createXLinkOut()
+        xout_rgb.setStreamName("rgb")
+        if use_hq:
+            self.nodes.cam_rgb.video.link(xout_rgb.input)
+        else:
+            self.nodes.cam_rgb.preview.link(xout_rgb.input)
 
-        def print_packet_info(packet, stream_name):
-            meta = packet.getMetadata()
-            print("[{:.6f} {:15s}]".format(time() - time_start, stream_name), end='')
-            if meta is not None:
-                source = meta.getCameraName()
-                if stream_name.startswith('disparity') or stream_name.startswith('depth'):
-                    source += '(rectif)'
-                print(" {:.6f}".format(meta.getTimestamp()), meta.getSequenceNum(), source, end='')
-                print('', meta.getFrameWidth(), meta.getFrameHeight(), meta.getFrameBytesPP(), end='')
-            print()
-            return
+    def create_depth(self, dct, median, lr):
+        self.nodes.stereo = self.p.createStereoDepth()
+        self.nodes.stereo.setOutputDepth(True)
+        self.nodes.stereo.setOutputRectified(True)
+        self.nodes.stereo.setConfidenceThreshold(dct)
+        self.nodes.stereo.setMedianFilter(median)
+        self.nodes.stereo.setLeftRightCheck(lr)
 
-        def keypress_handler(self, key, stream_names):
-            cams = ['rgb', 'mono']
-            self.cam_idx = getattr(self, 'cam_idx', 0)  # default: 'rgb'
-            cam = cams[self.cam_idx]
-            cam_c = depthai.CameraControl.CamId.RGB
-            cam_l = depthai.CameraControl.CamId.LEFT
-            cam_r = depthai.CameraControl.CamId.RIGHT
-            cmd_ae_region = depthai.CameraControl.Command.AE_REGION
-            cmd_exp_comp  = depthai.CameraControl.Command.EXPOSURE_COMPENSATION
-            cmd_set_focus = depthai.CameraControl.Command.MOVE_LENS
-            cmd_set_exp   = depthai.CameraControl.Command.AE_MANUAL
-            keypress_handler_lut = {
-                ord('f'): lambda: self.device.request_af_trigger(),
-                ord('1'): lambda: self.device.request_af_mode(depthai.AutofocusMode.AF_MODE_AUTO),
-                ord('2'): lambda: self.device.request_af_mode(depthai.AutofocusMode.AF_MODE_CONTINUOUS_VIDEO),
-                # 5,6,7,8,9,0: short example for using ISP 3A controls for Mono cameras
-                ord('5'): lambda: self.device.send_camera_control(cam_l, cmd_ae_region, '0 0 200 200 1'),
-                ord('6'): lambda: self.device.send_camera_control(cam_l, cmd_ae_region, '1000 0 200 200 1'),
-                ord('7'): lambda: self.device.send_camera_control(cam_l, cmd_exp_comp, '-2'),
-                ord('8'): lambda: self.device.send_camera_control(cam_l, cmd_exp_comp, '+2'),
-                ord('9'): lambda: self.device.send_camera_control(cam_r, cmd_exp_comp, '-2'),
-                ord('0'): lambda: self.device.send_camera_control(cam_r, cmd_exp_comp, '+2'),
-            }
-            if key in keypress_handler_lut:
-                keypress_handler_lut[key]()
-            elif key == ord('c'):
-                if 'jpegout' in stream_names:
-                    self.device.request_jpeg()
+        # Create mono left/right cameras if we haven't already
+        if not hasattr(self.nodes, 'mono_left'): self.create_left_cam(create_xout=False)
+        if not hasattr(self.nodes, 'mono_right'): self.create_right_cam(create_xout=True)
+
+        self.nodes.mono_left.out.link(self.nodes.stereo.left)
+        self.nodes.mono_right.out.link(self.nodes.stereo.right)
+
+        if self.nn_manager.sbb: # If we want to displaty SBBs, we need to send depth frames as well (via XLink)
+            self.nodes.xout_depth = self.p.createXLinkOut()
+            self.nodes.xout_depth.setStreamName("depth")
+            # If we don't want camera/NN in sync, set the disparity as the source for the XOut
+            if not conf.args.sync:
+                self.nodes.stereo.depth.link(self.nodes.xout_depth.input)
+
+    def create_left_cam(self, create_xout):
+        self.nodes.mono_left = self.p.createMonoCamera()
+        self.nodes.mono_left.setResolution(dai.MonoCameraProperties.SensorResolution.THE_400_P)
+        self.nodes.mono_left.setBoardSocket(dai.CameraBoardSocket.LEFT)
+
+        if create_xout:
+            self.nodes.xout_left = self.p.createXLinkOut()
+            self.nodes.xout_left.setStreamName("left")
+
+    def create_right_cam(self, create_xout):
+        self.nodes.mono_right = self.p.createMonoCamera()
+        self.nodes.mono_right.setResolution(dai.MonoCameraProperties.SensorResolution.THE_400_P)
+        self.nodes.mono_right.setBoardSocket(dai.CameraBoardSocket.RIGHT)
+
+        if create_xout:
+            self.nodes.xout_right = self.p.createXLinkOut()
+            self.nodes.xout_right.setStreamName("right")
+
+    def create_nn(self):
+        # If we want depth information , create ImageManip that will be the input for the NN
+        if conf.useDepth:
+            self.nodes.manip = self.p.createImageManip()
+            self.nodes.manip.initialConfig.setResize(in_w, in_h)
+            # The NN model expects BGR input. By default ImageManip output type would be same as input (gray in this case)
+            self.nodes.manip.initialConfig.setFrameType(dai.RawImgFrame.Type.BGR888p)
+            # Set the stereo node's rectifiedRight stream as the ImageManip's source
+            self.nodes.stereo.rectifiedRight.link(self.nodes.manip.inputImage)
+
+            # If we don't want camera/NN in sync, set the rectifiedRight as the source for the XOut
+            if not conf.args.sync:
+                self.nodes.manip.out.link(self.nodes.xout_right.input)
+
+        if callable(self.nn_manager.create_nn_pipeline):
+            self.nn_manager.create_nn_pipeline(self.p, self.nodes)
+
+nn_manager = NNetManager(
+    model_name=conf.getModelName(),
+    model_dir=conf.getModelDir(),
+    source=conf.getModelSource(),
+    use_depth=conf.useDepth,
+    use_hq=conf.useHQ
+)
+
+pm = PipelineManager(nn_manager)
+
+if conf.useDepth:
+    pm.create_depth(conf.args.disparity_confidence_threshold, median, conf.args.stereo_lr_check)
+elif conf.useCamera:
+    pm.create_color_cam(conf.useHQ)
+
+pm.create_nn()
+
+
+# Pipeline is defined, now we can connect to the device
+with dai.Device(pm.p) as device:
+    # Start pipeline
+    device.startPipeline()
+    nn_manager.addDevice(device)
+    if conf.useDepth:
+        q_right = device.getOutputQueue(name="right", maxSize=1, blocking=False)
+        fps = FPSHandler()
+    elif conf.useCamera:
+        q_rgb = device.getOutputQueue(name="rgb", maxSize=1, blocking=False)
+        fps = FPSHandler()
+    else:
+        cap = cv2.VideoCapture(conf.args.video)
+        fps = FPSHandler(cap)
+        seq_num = 0
+
+    frame = None
+    detections = []
+    # Spatial bounding box ROIs (region of interests)
+    color = (255, 255, 255)
+
+    while True:
+        fps.next_iter()
+        if conf.useDepth:
+            in_right = q_right.get()
+            frame = in_right.getCvFrame()
+            # Since rectified frames are horizontally flipped by default
+            # if not hq:
+            frame = cv2.flip(frame, 1)
+            fps.tick('right')
+
+            if nn_manager.sbb: # Get spatial bounding boxes and depth map
+                depth_frame = nn_manager.depth_out.get().getFrame()
+                depth_frame = cv2.normalize(depth_frame, None, 255, 0, cv2.NORM_INF, cv2.CV_8UC1)
+                depth_frame = cv2.equalizeHist(depth_frame)
+                depth_frame = cv2.applyColorMap(depth_frame, cv2.COLORMAP_TURBO)
+
+                sbb = nn_manager.sbb_out.tryGet()
+                sbb_rois = sbb.getConfigData() if sbb is not None else []
+                for roi_data in sbb_rois:
+                    roi = roi_data.roi
+                    roi = roi.denormalize(depth_frame.shape[1], depth_frame.shape[0])
+                    top_left = roi.topLeft()
+                    bottom_right = roi.bottomRight()
+                    # Display SBB on the disparity map
+                    cv2.rectangle(depth_frame, (int(top_left.x), int(top_left.y)), (int(bottom_right.x), int(bottom_right.y)), color, cv2.FONT_HERSHEY_SCRIPT_SIMPLEX)
+
+                cv2.imshow("disparity", depth_frame)
+
+        elif conf.useCamera:
+            in_rgb = q_rgb.get()
+            if in_rgb is not None:
+                if conf.useHQ:
+                    yuv = in_rgb.getData().reshape((in_rgb.getHeight() * 3 // 2, in_rgb.getWidth()))
+                    frame = cv2.cvtColor(yuv, cv2.COLOR_YUV2BGR_NV12)
                 else:
-                    print("'jpegout' stream not enabled. Try settings -s jpegout to enable it")
-            elif key == ord('s'):  # switch selected camera for manual exposure control
-                self.cam_idx = (self.cam_idx + 1) % len(cams)
-                print("======================= Current camera to control:", cams[self.cam_idx])
-            # RGB manual focus/exposure controls:
-            # Control:      key[dec/inc]  min..max
-            # exposure time:     i   o    1..33333 [us]
-            # sensitivity iso:   k   l    100..1600
-            # focus:             ,   .    0..255 [far..near]
-            elif key == ord('i') or key == ord('o') or key == ord('k') or key == ord('l'):
-                max_exp_us = int(1000*1000 / config['camera'][cam]['fps'])
-                self.rgb_exp = getattr(self, 'rgb_exp', 20000)  # initial
-                self.rgb_iso = getattr(self, 'rgb_iso', 800)  # initial
-                rgb_iso_step = 50
-                rgb_exp_step = max_exp_us // 20  # split in 20 steps
-                if key == ord('i'): self.rgb_exp -= rgb_exp_step
-                if key == ord('o'): self.rgb_exp += rgb_exp_step
-                if key == ord('k'): self.rgb_iso -= rgb_iso_step
-                if key == ord('l'): self.rgb_iso += rgb_iso_step
-                if self.rgb_exp < 1:     self.rgb_exp = 1
-                if self.rgb_exp > max_exp_us: self.rgb_exp = max_exp_us
-                if self.rgb_iso < 100:   self.rgb_iso = 100
-                if self.rgb_iso > 1600:  self.rgb_iso = 1600
-                print("===================================", cam, "set exposure:", self.rgb_exp, "iso:", self.rgb_iso)
-                exp_arg = str(self.rgb_exp) + ' ' + str(self.rgb_iso) + ' 33333'
-                if cam == 'rgb':
-                    self.device.send_camera_control(cam_c, cmd_set_exp, exp_arg)
-                elif cam == 'mono':
-                    self.device.send_camera_control(cam_l, cmd_set_exp, exp_arg)
-                    self.device.send_camera_control(cam_r, cmd_set_exp, exp_arg)
-            elif key == ord(',') or key == ord('.'):
-                self.rgb_manual_focus = getattr(self, 'rgb_manual_focus', 200)  # initial
-                rgb_focus_step = 3
-                if key == ord(','): self.rgb_manual_focus -= rgb_focus_step
-                if key == ord('.'): self.rgb_manual_focus += rgb_focus_step
-                if self.rgb_manual_focus < 0:   self.rgb_manual_focus = 0
-                if self.rgb_manual_focus > 255: self.rgb_manual_focus = 255
-                print("========================================== RGB set focus:", self.rgb_manual_focus)
-                focus_arg = str(self.rgb_manual_focus)
-                self.device.send_camera_control(cam_c, cmd_set_focus, focus_arg)
-            return
-
-        for stream in stream_names:
-            if stream in ["disparity", "disparity_color", "depth"]:
-                cv2.namedWindow(stream)
-                trackbar_name = 'Disparity confidence'
-                conf_thr_slider_min = 0
-                conf_thr_slider_max = 255
-                cv2.createTrackbar(trackbar_name, stream, conf_thr_slider_min, conf_thr_slider_max, self.on_trackbar_change)
-                cv2.setTrackbarPos(trackbar_name, stream, args['disparity_confidence_threshold'])
-
-        right_rectified = None
-        pcl_converter = None
-
-        ops = 0
-        prevTime = time()
-        if args['verbose']: print_packet_info_header()
-        while self.runThread:
-            # retreive data from the device
-            # data is stored in packets, there are nnet (Neural NETwork) packets which have additional functions for NNet result interpretation
-            self.nnet_packets, self.data_packets = p.get_available_nnet_and_data_packets(blocking=True)
-
-            ### Uncomment to print ops
-            # ops = ops + 1
-            # if time() - prevTime > 1.0:
-            #     print('OPS: ', ops)
-            #     ops = 0
-            #     prevTime = time()
-
-            packets_len = len(self.nnet_packets) + len(self.data_packets)
-            if packets_len != 0:
-                self.reset_process_wd()
-            else:
-                cur_time = monotonic()
-                if cur_time > wd_cutoff:
-                    print("process watchdog timeout")
-                    os._exit(10)
-
-            for _, nnet_packet in enumerate(self.nnet_packets):
-                if args['verbose']: print_packet_info(nnet_packet, 'NNet')
-
-                meta = nnet_packet.getMetadata()
-                camera = 'rgb'
-                if meta != None:
-                    camera = meta.getCameraName()
-                nnet_prev["nnet_source"][camera] = nnet_packet
-                nnet_prev["entries_prev"][camera] = decode_nn(nnet_packet, config=config, NN_json=NN_json)
-                frame_count['metaout'] += 1
-                frame_count['nn'][camera] += 1
-
-            for packet in self.data_packets:
-                window_name = packet.stream_name
-                if packet.stream_name not in stream_names:
-                    continue  # skip streams that were automatically added
-                if args['verbose']: print_packet_info(packet, packet.stream_name)
-                packetData = packet.getData()
-                if packetData is None:
-                    print('Invalid packet data!')
-                    continue
-                elif packet.stream_name == 'previewout':
-                    meta = packet.getMetadata()
-                    camera = 'rgb'
-                    if meta != None:
-                        camera = meta.getCameraName()
-
-                    window_name = 'previewout-' + camera
-                    # the format of previewout image is CHW (Chanel, Height, Width), but OpenCV needs HWC, so we
-                    # change shape (3, 300, 300) -> (300, 300, 3)
-                    data0 = packetData[0, :, :]
-                    data1 = packetData[1, :, :]
-                    data2 = packetData[2, :, :]
-                    frame = cv2.merge([data0, data1, data2])
-                    if nnet_prev["entries_prev"][camera] is not None:
-                        frame = show_nn(nnet_prev["entries_prev"][camera], frame, NN_json=NN_json, config=config)
-                        if enable_object_tracker and tracklets is not None:
-                            frame = show_tracklets(tracklets, frame, labels)
-                    cv2.putText(frame, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 0))
-                    cv2.putText(frame, "NN fps: " + str(frame_count_prev['nn'][camera]), (2, frame.shape[0] - 4), cv2.FONT_HERSHEY_SIMPLEX, 0.4, (0, 255, 0))
-                    cv2.imshow(window_name, frame)
-                elif packet.stream_name in ['left', 'right', 'disparity', 'rectified_left', 'rectified_right']:
-                    frame_bgr = packetData
-                    if args['pointcloud'] and packet.stream_name == 'rectified_right':
-                        right_rectified = packetData
-                    cv2.putText(frame_bgr, packet.stream_name, (25, 25), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 0))
-                    cv2.putText(frame_bgr, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 0))
-                    if args['draw_bb_depth']:
-                        camera = args['cnn_camera']
-                        if packet.stream_name == 'disparity':
-                            if camera == 'left_right':
-                                camera = 'right'
-                        elif camera != 'rgb':
-                            camera = packet.getMetadata().getCameraName()
-                        if nnet_prev["entries_prev"][camera] is not None:
-                            frame_bgr = show_nn(nnet_prev["entries_prev"][camera], frame_bgr, NN_json=NN_json, config=config, nn2depth=nn2depth)
-                    cv2.imshow(window_name, frame_bgr)
-                elif packet.stream_name.startswith('depth') or packet.stream_name == 'disparity_color':
-                    frame = packetData
-
-                    if len(frame.shape) == 2:
-                        if frame.dtype == np.uint8:  # grayscale
-                            cv2.putText(frame, packet.stream_name, (25, 25), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 255))
-                            cv2.putText(frame, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 255))
-                        else:  # uint16
-                            if args['pointcloud'] and "depth" in stream_names and "rectified_right" in stream_names and right_rectified is not None:
-                                try:
-                                    from depthai_helpers.projector_3d import PointCloudVisualizer
-                                except ImportError as e:
-                                    raise ImportError(f"\033[1;5;31mError occured when importing PCL projector: {e} \033[0m ")
-                                if pcl_converter is None:
-                                    pcl_converter = PointCloudVisualizer(self.device.get_right_intrinsic(), 1280, 720)
-                                right_rectified = cv2.flip(right_rectified, 1)
-                                pcl_converter.rgbd_to_projection(frame, right_rectified)
-                                pcl_converter.visualize_pcd()
-
-                            frame = (65535 // frame).astype(np.uint8)
-                            # colorize depth map, comment out code below to obtain grayscale
-                            frame = cv2.applyColorMap(frame, cv2.COLORMAP_HOT)
-                            # frame = cv2.applyColorMap(frame, cv2.COLORMAP_JET)
-                            cv2.putText(frame, packet.stream_name, (25, 25), cv2.FONT_HERSHEY_SIMPLEX, 1.0, 255)
-                            cv2.putText(frame, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, 255)
-                    else:  # bgr
-                        cv2.putText(frame, packet.stream_name, (25, 25), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (255, 255, 255))
-                        cv2.putText(frame, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, 255)
-
-                    if args['draw_bb_depth']:
-                        camera = args['cnn_camera']
-                        if camera == 'left_right':
-                            camera = 'right'
-                        if nnet_prev["entries_prev"][camera] is not None:
-                            frame = show_nn(nnet_prev["entries_prev"][camera], frame, NN_json=NN_json, config=config, nn2depth=nn2depth)
-                    cv2.imshow(window_name, frame)
-
-                elif packet.stream_name == 'jpegout':
-                    jpg = packetData
-                    mat = cv2.imdecode(jpg, cv2.IMREAD_COLOR)
-                    cv2.imshow('jpegout', mat)
-
-                elif packet.stream_name == 'video':
-                    videoFrame = packetData
-                    videoFrame.tofile(video_file)
-                    # mjpeg = packetData
-                    # mat = cv2.imdecode(mjpeg, cv2.IMREAD_COLOR)
-                    # cv2.imshow('mjpeg', mat)
-                elif packet.stream_name == 'color':
-                    meta = packet.getMetadata()
-                    w = meta.getFrameWidth()
-                    h = meta.getFrameHeight()
-                    yuv420p = packetData.reshape((h * 3 // 2, w))
-                    bgr = cv2.cvtColor(yuv420p, cv2.COLOR_YUV2BGR_IYUV)
-                    scale = configMan.getColorPreviewScale()
-                    bgr = cv2.resize(bgr, (int(w * scale), int(h * scale)), interpolation=cv2.INTER_AREA)
-                    cv2.putText(bgr, packet.stream_name, (25, 25), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 0))
-                    cv2.putText(bgr, "fps: " + str(frame_count_prev[window_name]), (25, 50), cv2.FONT_HERSHEY_SIMPLEX, 1.0, (0, 0, 0))
-                    cv2.imshow("color", bgr)
-
-                elif packet.stream_name == 'meta_d2h':
-                    str_ = packet.getDataAsStr()
-                    dict_ = json.loads(str_)
-
-                    print('meta_d2h Temp',
-                          ' CSS:' + '{:6.2f}'.format(dict_['sensors']['temperature']['css']),
-                          ' MSS:' + '{:6.2f}'.format(dict_['sensors']['temperature']['mss']),
-                          ' UPA:' + '{:6.2f}'.format(dict_['sensors']['temperature']['upa0']),
-                          ' DSS:' + '{:6.2f}'.format(dict_['sensors']['temperature']['upa1']))
-                elif packet.stream_name == 'object_tracker':
-                    tracklets = packet.getObjectTracker()
-
-                frame_count[window_name] += 1
-
-            t_curr = time()
-            if t_start + 1.0 < t_curr:
-                t_start = t_curr
-                # print("metaout fps: " + str(frame_count_prev["metaout"]))
-
-                stream_windows = []
-                for s in stream_names:
-                    if s == 'previewout':
-                        for cam in NN_cams:
-                            stream_windows.append(s + '-' + cam)
-                            frame_count_prev['nn'][cam] = frame_count['nn'][cam]
-                            frame_count['nn'][cam] = 0
-                    else:
-                        stream_windows.append(s)
-                for w in stream_windows:
-                    frame_count_prev[w] = frame_count[w]
-                    frame_count[w] = 0
-
-            key = cv2.waitKey(1)
-            if key == ord('q'):
+                    frame_data = in_rgb.getData().reshape(3, in_rgb.getHeight(), in_rgb.getWidth())
+                    frame = np.ascontiguousarray(frame_data.transpose(1, 2, 0))
+                fps.tick('rgb')
+        else:
+            read_correctly, vid_frame = cap.read()
+            if not read_correctly:
                 break
-            else:
-                keypress_handler(self, key, stream_names)
 
-        del p  # in order to stop the pipeline object should be deleted, otherwise device will continue working. This is required if you are going to add code after the main loop, otherwise you can ommit it.
-        del self.device
-        cv2.destroyAllWindows()
+            scaled_frame = cv2.resize(vid_frame, (in_w, in_h))
+            frame_nn = dai.ImgFrame()
+            frame_nn.setSequenceNum(seq_num)
+            frame_nn.setWidth(in_w)
+            frame_nn.setHeight(in_h)
+            frame_nn.setData(to_planar(scaled_frame))
+            nn_manager.input.send(frame_nn)
+            seq_num += 1
 
-        # Close video output file if was opened
-        if video_file is not None:
-            video_file.close()
+            # if high quality, send original frames
+            frame = vid_frame if conf.useHQ else scaled_frame
+            fps.tick('rgb')
 
-        print('py: DONE.')
+        in_nn = nn_manager.output.tryGetAll()
+        if len(in_nn) > 0:
+            if nn_manager.output_format == "detection":
+                detections = in_nn[-1].detections
+            for packet in in_nn:
+                if nn_manager.output_format is None:
+                    try:
+                        print("Received NN packet: ", to_tensor_result(packet))
+                    except Exception as ex:
+                        print("Received NN packet: <Preview unabailable: {}>".format(ex))
+                fps.tick('nn')
 
+        if frame is not None:
+            # Scale the frame by --scale factor
+            if not conf.args.scale == 1.0:
+                h, w, c = frame.shape
+                frame = cv2.resize(frame, (int(w * conf.args.scale), int(h * conf.args.scale)), interpolation=cv2.INTER_AREA)
 
-if __name__ == "__main__":
-    dai = DepthAI()
-    dai.startLoop()
+            # if the frame is available, draw bounding boxes on it and show the frame
+            for detection in detections:
+                if conf.useDepth: # Since rectified frames are horizontally flipped by default
+                    swap = detection.xmin
+                    detection.xmin = 1 - detection.xmax
+                    detection.xmax = 1 - swap
+
+                bbox = frame_norm(frame, [detection.xmin, detection.ymin, detection.xmax, detection.ymax])
+                cv2.rectangle(frame, (bbox[0], bbox[1]), (bbox[2], bbox[3]), (255, 0, 0), 2)
+                cv2.putText(frame, nn_manager.get_label_text(detection.label), (bbox[0] + 10, bbox[1] + 20), cv2.FONT_HERSHEY_TRIPLEX, 0.5, 255)
+                cv2.putText(frame, f"{int(detection.confidence * 100)}%", (bbox[0] + 10, bbox[1] + 40), cv2.FONT_HERSHEY_TRIPLEX, 0.5, 255)
+
+                if conf.useDepth: # Display coordinates as well
+                    cv2.putText(frame, f"X: {int(detection.spatialCoordinates.x)} mm", (bbox[0] + 10, bbox[1] + 60), cv2.FONT_HERSHEY_TRIPLEX, 0.5, color)
+                    cv2.putText(frame, f"Y: {int(detection.spatialCoordinates.y)} mm", (bbox[0] + 10, bbox[1] + 75), cv2.FONT_HERSHEY_TRIPLEX, 0.5, color)
+                    cv2.putText(frame, f"Z: {int(detection.spatialCoordinates.z)} mm", (bbox[0] + 10, bbox[1] + 90), cv2.FONT_HERSHEY_TRIPLEX, 0.5, color)
+
+            frame_fps = f"RIGHT FPS: {round(fps.tick_fps('right'), 1)}" if conf.useDepth else f"RGB FPS: {round(fps.tick_fps('rgb'), 1)}"
+            cv2.putText(frame, frame_fps, (5, 15), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0))
+
+            cv2.putText(frame, f"NN FPS:  {round(fps.tick_fps('nn'), 1)}", (5, 30), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 255, 0))
+            cv2.imshow(conf.getModelSource(), frame)
+
+        if cv2.waitKey(1) == ord('q'):
+            break
