+++ depthai_helpers/config_manager.py
@@ -1,30 +1,102 @@
 import os
-import json
 import platform
 import subprocess
-from pathlib import Path
-import consts.resource_paths
 import urllib.request
+from pathlib import Path
+import depthai as dai
 
-from depthai_helpers import utils
-from model_compiler.model_compiler import download_and_compile_NN_model
-from consts.resource_paths import nn_resource_path as model_zoo_folder
+import blobconverter
+
 from depthai_helpers.cli_utils import cli_print, PrintColors
 
-class DepthConfigManager:
+default_input_dims = {
+    # TODO remove once fetching input size from nn blob is possible
+    "mobilenet-ssd": "300x300",
+    "face-detection-adas-0001": "672x384",
+    "face-detection-retail-0004": "300x300",
+    "pedestrian-detection-adas-0002": "672x384",
+    "person-detection-retail-0013": "544x320",
+    "person-vehicle-bike-detection-crossroad-1016": "512x512",
+    "vehicle-detection-adas-0002": "672x384",
+    "vehicle-license-plate-detection-barrier-0106": "300x300",
+    "tiny-yolo-v3": "416x416",
+    "yolo-v3": "416x416"
+}
+DEPTHAI_ZOO = Path(__file__).parent.parent / Path(f"resources/nn/")
+
+
+class ConfigManager:
     labels = ""
     NN_config = None
     custom_fw_commit = ''
 
     def __init__(self, args):
         self.args = args
-        self.stream_list = args['streams']
-        self.calc_dist_to_bb = not self.args['disable_depth']
 
-        # Prepare handler methods (decode_nn, show_nn) for the network we want to run.
-        self.importAndSetCallbacksForNN()
-        self.generateJsonConfig()
+    @property
+    def debug(self):
+        return not self.args.no_debug
 
+    @property
+    def useCamera(self):
+        return not self.args.video
+
+    @property
+    def useHQ(self):
+        return self.args.high_quality
+
+    @property
+    def useDepth(self):
+        return not self.args.disable_depth
+
+    def getModelSource(self):
+        if self.useDepth:
+            return "right"
+        elif self.useCamera:
+            return "rgb"
+        else:
+            return "host"
+
+    def getModelName(self):
+        if self.args.cnn_model:
+            return self.args.cnn_model
+        model_dir = self.getModelDir()
+        if model_dir is not None:
+            return Path(model_dir).stem
+
+    def getModelDir(self):
+        if self.args.cnn_path:
+            return self.args.cnn_path
+        if self.args.cnn_model is not None and (DEPTHAI_ZOO / self.args.cnn_model).exists():
+            return DEPTHAI_ZOO / self.args.cnn_model
+        print("dupa")
+
+    def getInputSize(self):
+        if self.args.cnn_input_size is None:
+            if self.args.cnn_model not in default_input_dims:
+                raise RuntimeError(
+                    "Unable to determine the nn input size. Please use --cnn_input_size flag to specify it in WxW format: -nn-size <width>x<height>")
+            return map(int, default_input_dims[self.args.cnn_model].split('x'))
+        else:
+            return map(int, self.args.cnn_input_size.split('x'))
+
+    def getRgbResolution(self):
+        if self.args.rgb_resolution == 2160:
+            return dai.ColorCameraProperties.SensorResolution.THE_4_K
+        elif self.args.rgb_resolution == 3040:
+            return dai.ColorCameraProperties.SensorResolution.THE_12_MP
+        else:
+            return dai.ColorCameraProperties.SensorResolution.THE_1080_P
+
+    def getMedianFilter(self):
+        if self.args.stereo_median_size == 3:
+            return dai.StereoDepthProperties.MedianFilter.KERNEL_3x3
+        elif self.args.stereo_median_size == 5:
+            return dai.StereoDepthProperties.MedianFilter.KERNEL_5x5
+        elif self.args.stereo_median_size == 7:
+            return dai.StereoDepthProperties.MedianFilter.KERNEL_7x7
+        else:
+            return dai.StereoDepthProperties.MedianFilter.MEDIAN_OFF
 
     def getUsb2Mode(self):
         usb2_mode = False
@@ -34,12 +106,6 @@ class DepthConfigManager:
         else:
             usb2_mode = False
         return usb2_mode
-
-    def getColorPreviewScale(self):
-        if self.args['color_scale']:
-            return self.args['color_scale']
-        else:
-            return 1.0
 
     def getCustomFirmwarePath(self, commit):
         fwdir = '.fw_cache/'
@@ -86,65 +152,6 @@ class DepthConfigManager:
 
         return cmd_file, debug_mode
 
-
-    def importAndSetCallbacksForNN(self):
-        # why inline imports? Could just make a dict for all these to make managing them easier and less verbose.
-        from depthai_helpers.mobilenet_ssd_handler import decode_mobilenet_ssd, show_mobilenet_ssd, decode_mobilenet_ssd_json
-        self.decode_nn=decode_mobilenet_ssd
-        self.show_nn=show_mobilenet_ssd
-        self.decode_nn_json=decode_mobilenet_ssd_json
-
-        if self.args['cnn_model'] == 'age-gender-recognition-retail-0013':
-            from depthai_helpers.age_gender_recognition_handler import decode_age_gender_recognition, show_age_gender_recognition, decode_age_gender_recognition_json
-            self.decode_nn=decode_age_gender_recognition
-            self.show_nn=show_age_gender_recognition
-            self.decode_nn_json=decode_age_gender_recognition_json
-            self.calc_dist_to_bb=False
-
-        if self.args['cnn_model'] == 'emotions-recognition-retail-0003':
-            from depthai_helpers.emotion_recognition_handler import decode_emotion_recognition, show_emotion_recognition, decode_emotion_recognition_json
-            self.decode_nn=decode_emotion_recognition
-            self.show_nn=show_emotion_recognition
-            self.decode_nn_json=decode_emotion_recognition_json
-            self.calc_dist_to_bb=False
-
-        if self.args['cnn_model'] in ['tiny-yolo-v3', 'yolo-v3']:
-            from depthai_helpers.tiny_yolo_v3_handler import decode_tiny_yolo, show_tiny_yolo, decode_tiny_yolo_json
-            self.decode_nn=decode_tiny_yolo
-            self.show_nn=show_tiny_yolo
-            self.decode_nn_json=decode_tiny_yolo_json
-
-        if self.args['cnn_model'] in ['facial-landmarks-35-adas-0002', 'landmarks-regression-retail-0009']:
-            from depthai_helpers.landmarks_recognition_handler import decode_landmarks_recognition, show_landmarks_recognition, decode_landmarks_recognition_json
-            self.decode_nn=decode_landmarks_recognition
-            self.show_nn=show_landmarks_recognition
-            self.decode_nn_json=decode_landmarks_recognition_json
-            self.calc_dist_to_bb=False
-
-        # backward compatibility
-        if self.args['cnn_model'] == 'openpose':
-            self.args['cnn_model'] = 'human-pose-estimation-0001'
-        if self.args['cnn_model'] == 'human-pose-estimation-0001':
-            from depthai_helpers.openpose_handler import decode_openpose, show_openpose
-            self.decode_nn=decode_openpose
-            self.show_nn=show_openpose
-            self.calc_dist_to_bb=False
-            
-        if self.args['cnn_model'] == 'openpose2':
-            self.args['cnn_model'] = 'mobileNetV2-PoseEstimation'
-        if self.args['cnn_model'] == 'mobileNetV2-PoseEstimation':
-            from depthai_helpers.openpose2_handler import decode_openpose, show_openpose
-            self.decode_nn=decode_openpose
-            self.show_nn=show_openpose
-            self.calc_dist_to_bb=False
-
-        if self.args['cnn_model'] == 'deeplabv3p_person':
-            from depthai_helpers.deeplabv3p_person import decode_deeplabv3p, show_deeplabv3p
-            self.decode_nn=decode_deeplabv3p
-            self.show_nn=show_deeplabv3p
-            self.calc_dist_to_bb=False
-
-
     def linuxCheckApplyUsbRules(self):
         if platform.system() == 'Linux':
             ret = subprocess.call(['grep', '-irn', 'ATTRS{idVendor}=="03e7"', '/etc/udev/rules.d'], stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL)
@@ -156,224 +163,39 @@ class DepthConfigManager:
                 "Disconnect/connect usb cable on host! \n", PrintColors.RED)
                 os._exit(1)
 
-    def getMaxShaveNumbers(self):
-        stream_names = [stream if isinstance(stream, str) else stream['name'] for stream in self.stream_list]
-        max_shaves = 14
-        if self.args['rgb_resolution'] != 1080:
-            max_shaves = 11
-            if 'object_tracker' in stream_names:
-                max_shaves = 9
-        elif 'object_tracker' in stream_names:
-            max_shaves = 12
-
-        return max_shaves
-
-    def generateJsonConfig(self):
-
-        # something to verify usb rules are good?
-        self.linuxCheckApplyUsbRules()
-
-        max_shave_nr = self.getMaxShaveNumbers()
-        shave_nr = max_shave_nr if self.args['shaves'] is None else self.args['shaves']
-        cmx_slices = shave_nr if self.args['cmx_slices'] is None else self.args['cmx_slices']
-        NCE_nr = 1 if self.args['NN_engines'] is None else self.args['NN_engines']
-
-        # left_right double NN check.
-        if self.args['cnn_camera'] in ['left_right', 'rectified_left_right']:
-            if NCE_nr != 2:
-                cli_print('Running NN on both cams requires 2 NN engines!', PrintColors.RED)
-                NCE_nr = 2
-
-        if NCE_nr == 2:
-            shave_nr = shave_nr - (shave_nr % 2)
-            cmx_slices = cmx_slices - (cmx_slices % 2)
-
-        # Get blob files
-        blobMan = BlobManager(self.args, self.calc_dist_to_bb, shave_nr, cmx_slices, NCE_nr)
-        self.NN_config = blobMan.getNNConfig()
-        try:
-            self.labels = self.NN_config['mappings']['labels']
-        except:
-            self.labels = None
-            print("Labels not found in json!")
-
-        try:
-            output_format = self.NN_config['NN_config']['output_format']
-        except:
-            NN_json = {}
-            NN_json['NN_config'] = {}
-            NN_json['NN_config']['output_format'] = "raw"
-            self.NN_config = NN_json
-            output_format = "raw"
-
-        if output_format == "raw" and self.calc_dist_to_bb == True:
-            cli_print("WARNING: Depth calculation with raw output format is not supported! It's only supported for YOLO/mobilenet based NNs, disabling calc_dist_to_bb", PrintColors.WARNING)
-            self.calc_dist_to_bb = False
-        
-        stream_names = [stream if isinstance(stream, str) else stream['name'] for stream in self.stream_list]
-        if ('disparity' in stream_names or 'disparity_color' in stream_names) and self.calc_dist_to_bb == True:
-            cli_print("WARNING: Depth calculation with disparity/disparity_color streams is not supported! Disabling calc_dist_to_bb", PrintColors.WARNING)
-            self.calc_dist_to_bb = False
-
-        # check for known bad configurations
-        if 'depth' in stream_names and ('disparity' in stream_names or 'disparity_color' in stream_names):
-            print('ERROR: depth is mutually exclusive with disparity/disparity_color')
-            exit(2)
-
-        if self.args['stereo_lr_check'] == True:
-            raise ValueError("Left-right check option is still under development. Don;t enable it.")
-
-        # Do not modify the default values in the config Dict below directly. Instead, use the `-co` argument when running this script.
-        config = {
-            # Possible streams:
-            # ['left', 'right', 'jpegout', 'video', 'previewout', 'metaout', 'depth_sipp', 'disparity', 'depth_color_h']
-            # If "left" is used, it must be in the first position.
-            # To test depth use:
-            # 'streams': [{'name': 'depth_sipp', "max_fps": 12.0}, {'name': 'previewout', "max_fps": 12.0}, ],
-            'streams': self.args['streams'],
-            'depth':
-            {
-                'calibration_file': consts.resource_paths.calib_fpath,
-                'left_mesh_file': consts.resource_paths.left_mesh_fpath,
-                'right_mesh_file': consts.resource_paths.right_mesh_fpath,
-                'padding_factor': 0.3,
-                'depth_limit_m': 10.0, # In meters, for filtering purpose during x,y,z calc
-                'median_kernel_size': self.args['stereo_median_size'],
-                'lr_check': self.args['stereo_lr_check'],
-                'warp_rectify':
-                {
-                    'use_mesh' : self.args['use_mesh'], # if False, will use homography
-                    'mirror_frame': self.args['mirror_rectified'] == 'true', # if False, the disparity will be mirrored instead
-                    'edge_fill_color': 0, # gray 0..255, or -1 to replicate pixel values
-                },
-            },
-            'ai':
-            {
-                'blob_file': blobMan.blob_file,
-                'blob_file_config': blobMan.blob_file_config,
-                'blob_file2': blobMan.blob_file2,
-                'blob_file_config2': blobMan.blob_file_config2,
-                'calc_dist_to_bb': self.calc_dist_to_bb,
-                'keep_aspect_ratio': not self.args['full_fov_nn'],
-                'camera_input': self.args['cnn_camera'],
-                'shaves' : shave_nr,
-                'cmx_slices' : cmx_slices,
-                'NN_engines' : NCE_nr,
-            },
-            # object tracker
-            'ot':
-            {
-                'max_tracklets'        : 20, #maximum 20 is supported
-                'confidence_threshold' : 0.5, #object is tracked only for detections over this threshold
-            },
-            'board_config':
-            {
-                'swap_left_and_right_cameras': self.args['swap_lr'], # True for 1097 (RPi Compute) and 1098OBC (USB w/onboard cameras)
-                'left_fov_deg': self.args['field_of_view'], # Same on 1097 and 1098OBC
-                'rgb_fov_deg': self.args['rgb_field_of_view'],
-                'left_to_right_distance_cm': self.args['baseline'], # Distance between stereo cameras
-                'left_to_rgb_distance_cm': self.args['rgb_baseline'], # Currently unused
-                'store_to_eeprom': self.args['store_eeprom'],
-                'clear_eeprom': self.args['clear_eeprom'],
-                'override_eeprom': self.args['override_eeprom'],
-            },
-            'camera':
-            {
-                'rgb':
-                {
-                    # 3840x2160, 1920x1080
-                    # only UHD/1080p/30 fps supported for now
-                    'resolution_h': self.args['rgb_resolution'],
-                    'fps': self.args['rgb_fps'],
-                },
-                'mono':
-                {
-                    # 1280x720, 1280x800, 640x400 (binning enabled)
-                    'resolution_h': self.args['mono_resolution'],
-                    'fps': self.args['mono_fps'],
-                },
-            },
-            'app':
-            {
-                'sync_video_meta_streams': self.args['sync_video_meta'],
-                'sync_sequence_numbers'  : self.args['sync_sequence_numbers'],
-                'usb_chunk_KiB' : self.args['usb_chunk_KiB'],
-            },
-            #'video_config':
-            #{
-            #    'rateCtrlMode': 'cbr', # Options: cbr / vbr
-            #    'profile': 'h265_main', # Options: 'h264_baseline' / 'h264_main' / 'h264_high' / 'h265_main / 'mjpeg' '
-            #    'bitrate': 8000000, # When using CBR (H264/H265 only)
-            #    'maxBitrate': 8000000, # When using CBR (H264/H265 only)
-            #    'keyframeFrequency': 30, (H264/H265 only)
-            #    'numBFrames': 0, (H264/H265 only)
-            #    'quality': 80 # (0 - 100%) When using VBR or MJPEG profile
-            #}
-            #'video_config':
-            #{
-            #    'profile': 'mjpeg',
-            #    'quality': 95
-            #}
-        }
-
-        self.jsonConfig = self.postProcessJsonConfig(config)
-
-
-    def postProcessJsonConfig(self, config):
-        # merge board config, if exists
-        if self.args['board']:
-            board_path = Path(self.args['board'])
-            if not board_path.exists():
-                board_path = Path(consts.resource_paths.boards_dir_path) / Path(self.args['board'].upper()).with_suffix('.json')
-                if not board_path.exists():
-                    print('ERROR: Board config not found: {}'.format(board_path))
-                    os._exit(2)
-            with open(board_path) as fp:
-                board_config = json.load(fp)
-            utils.merge(board_config, config)
-
-        # handle config overwrite option.
-        if self.args['config_overwrite']:
-            self.args['config_overwrite'] = json.loads(self.args['config_overwrite'])
-            config = utils.merge(self.args['config_overwrite'],config)
-            print("Merged Pipeline config with overwrite",config)
-
-        # Append video stream if video recording was requested and stream is not already specified
-        self.video_file = None
-        if self.args['video'] is not None:
-            
-            # open video file
-            try:
-                self.video_file = open(self.args['video'], 'wb')
-                if config['streams'].count('video') == 0:
-                    config['streams'].append('video')
-            except IOError:
-                print("Error: couldn't open video file for writing. Disabled video output stream")
-                if config['streams'].count('video') == 1:
-                    config['streams'].remove('video')
-
-        return config
-
 
 class BlobManager:
-    def __init__(self, model_dir):
-        self.model_dir = Path(model_dir)
-        self.zoo_dir = self.model_dir.parent
-        self.model_name = self.model_dir.name
+    def __init__(self, model_name=None, model_dir=None):
+        self.model_dir = None
+        self.zoo_dir = None
+        self.config_file = None
+        self.use_zoo = False
+        if model_dir is None:
+            self.model_name = model_name
+            self.use_zoo = True
+        else:
+            self.model_dir = Path(model_dir)
+            self.zoo_dir = self.model_dir.parent
+            self.model_name = model_name or self.model_dir.name
+            self.config_file = self.model_dir / "model.yml"
+            if not self.config_file.exists():
+                self.use_zoo = True
+
         self.blob_path = None
 
     def compile(self, shaves, target='auto'):
-        sh, nce = (shaves // 2, 2) if shaves > 8 else (shaves, 1)
-        cmx = sh  # number of cmx slices should match the number of shaves
-        blob_path = self.model_dir / Path(self.model_name).with_suffix(f".blob.sh{shaves}")
-
-        if Path(blob_path).exists():
-            cli_print(f"Compiled model found, compiled for {shaves} shaves", PrintColors.GREEN)
+        if self.use_zoo:
+            return blobconverter.from_zoo(name=self.model_name, shaves=shaves)
         else:
-            cli_print(f"Compiling model for {shaves} shaves", PrintColors.RED)
-            ret = download_and_compile_NN_model(self.model_name, self.zoo_dir, sh, cmx, nce, blob_path, target)
-            if ret != 0:
-                cli_print("Model compile failed. Falling back to default.", PrintColors.WARNING)
-                raise RuntimeError("Model compilation failed! Not connected to the internet?")
-
-        return blob_path
+            return blobconverter.compile_blob(
+                blob_name=self.model_name,
+                req_data={
+                    "name": self.model_name,
+                    "use_zoo": True,
+                },
+                req_files={
+                    'config': self.config_file,
+                },
+                data_type="FP16",
+                shaves=shaves,
+            )
