--- depthai_helpers/openpose_handler.py
@@ -1,178 +1,0 @@
-import cv2
-import numpy as np
-
-keypointsMapping = ['Nose', 'Neck', 'R-Sho', 'R-Elb', 'R-Wr', 'L-Sho', 'L-Elb', 'L-Wr', 'R-Hip', 'R-Knee', 'R-Ank', 'L-Hip', 'L-Knee', 'L-Ank', 'R-Eye', 'L-Eye', 'R-Ear', 'L-Ear']
-POSE_PAIRS = [[1,2], [1,5], [2,3], [3,4], [5,6], [6,7], [1,8], [8,9], [9,10], [1,11], [11,12], [12,13], [1,0], [0,14], [14,16], [0,15], [15,17], [2,17], [5,16]]
-mapIdx = [[31,32], [39,40], [33,34], [35,36], [41,42], [43,44], [19,20], [21,22], [23,24], [25,26], [27,28], [29,30], [47,48], [49,50], [53,54], [51,52], [55,56], [37,38], [45,46]]
-colors = [[0,100,255], [0,100,255], [0,255,255], [0,100,255], [0,255,255], [0,100,255], [0,255,0], [255,200,100], [255,0,255], [0,255,0], [255,200,100], [255,0,255], [0,0,255], [255,0,0], [200,200,0], [255,0,0], [200,200,0], [0,0,0]]
-
-def getKeypoints(probMap, threshold=0.2):
-
-    mapSmooth = cv2.GaussianBlur(probMap, (3, 3), 0, 0)
-    mapMask = np.uint8(mapSmooth>threshold)
-    keypoints = []
-    contours = None
-    try:
-        #OpenCV4.x
-        contours, _ = cv2.findContours(mapMask, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-    except:
-        #OpenCV3.x
-        _, contours, _ = cv2.findContours(mapMask, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
-
-    for cnt in contours:
-        blobMask = np.zeros(mapMask.shape)
-        blobMask = cv2.fillConvexPoly(blobMask, cnt, 1)
-        maskedProbMap = mapSmooth * blobMask
-        _, maxVal, _, maxLoc = cv2.minMaxLoc(maskedProbMap)
-        keypoints.append(maxLoc + (probMap[maxLoc[1], maxLoc[0]],))
-
-    return keypoints
-
-
-def getValidPairs(outputs, w, h,detected_keypoints):
-    valid_pairs = []
-    invalid_pairs = []
-    n_interp_samples = 10
-    paf_score_th = 0.2
-    conf_th = 0.4
-    for k in range(len(mapIdx)):
-        
-        pafA = outputs[0, mapIdx[k][0], :, :]
-        pafB = outputs[0, mapIdx[k][1], :, :]
-        pafA = cv2.resize(pafA, (w, h))
-        pafB = cv2.resize(pafB, (w, h))
-        candA = detected_keypoints[POSE_PAIRS[k][0]]
-        
-        candB = detected_keypoints[POSE_PAIRS[k][1]]
-
-        nA = len(candA)
-        nB = len(candB)
-
-        if( nA != 0 and nB != 0):
-            valid_pair = np.zeros((0,3))
-            for i in range(nA):
-                max_j=-1
-                maxScore = -1
-                found = 0
-                for j in range(nB):
-                    d_ij = np.subtract(candB[j][:2], candA[i][:2])
-                    norm = np.linalg.norm(d_ij)
-                    if norm:
-                        d_ij = d_ij / norm
-                    else:
-                        continue
-                    interp_coord = list(zip(np.linspace(candA[i][0], candB[j][0], num=n_interp_samples),
-                                            np.linspace(candA[i][1], candB[j][1], num=n_interp_samples)))
-                    paf_interp = []
-                    for k in range(len(interp_coord)):
-                        paf_interp.append([pafA[int(round(interp_coord[k][1])), int(round(interp_coord[k][0]))],
-                                           pafB[int(round(interp_coord[k][1])), int(round(interp_coord[k][0]))] ])
-                    paf_scores = np.dot(paf_interp, d_ij)
-                    avg_paf_score = sum(paf_scores)/len(paf_scores)
-
-                    if ( len(np.where(paf_scores > paf_score_th)[0]) / n_interp_samples ) > conf_th :
-                        if avg_paf_score > maxScore:
-                            max_j = j
-                            maxScore = avg_paf_score
-                            found = 1
-                if found:
-                    valid_pair = np.append(valid_pair, [[candA[i][3], candB[max_j][3], maxScore]], axis=0)
-
-            valid_pairs.append(valid_pair)
-        else:
-            invalid_pairs.append(k)
-            valid_pairs.append([])
-    return valid_pairs, invalid_pairs
-
-
-def getPersonwiseKeypoints(valid_pairs, invalid_pairs,keypoints_list):
-    personwiseKeypoints = -1 * np.ones((0, 19))
-
-    for k in range(len(mapIdx)):
-        if k not in invalid_pairs:
-            partAs = valid_pairs[k][:,0]
-            partBs = valid_pairs[k][:,1]
-            indexA, indexB = np.array(POSE_PAIRS[k])
-
-            for i in range(len(valid_pairs[k])):
-                found = 0
-                person_idx = -1
-                for j in range(len(personwiseKeypoints)):
-                    if personwiseKeypoints[j][indexA] == partAs[i]:
-                        person_idx = j
-                        found = 1
-                        break
-
-                if found:
-                    personwiseKeypoints[person_idx][indexB] = partBs[i]
-                    personwiseKeypoints[person_idx][-1] += keypoints_list[partBs[i].astype(int), 2] + valid_pairs[k][i][2]
-
-                elif not found and k < 17:
-                    row = -1 * np.ones(19)
-                    row[indexA] = partAs[i]
-                    row[indexB] = partBs[i]
-                    row[-1] = sum(keypoints_list[valid_pairs[k][i,:2].astype(int), 2]) + valid_pairs[k][i][2]
-                    personwiseKeypoints = np.vstack([personwiseKeypoints, row])
-    return personwiseKeypoints
-    
-    
-
-threshold = 0.3
-nPoints = 18
-w=456
-h=256
-detected_keypoints = []
-def decode_openpose(nnet_packet, **kwargs):
-    # openpose from openvino has pafs in first output, keypoints in second
-    output_list = nnet_packet.getOutputsList()
-    heatmaps = np.copy(output_list[1])
-    pafs = np.copy(output_list[0])
-    
-    heatmaps = heatmaps.astype('float32')
-    pafs = pafs.astype('float32')
-    
-    outputs = np.concatenate((heatmaps,pafs),axis=1)
-
-    detected_keypoints = []
-    keypoints_list = np.zeros((0, 3))
-    keypoint_id = 0
-
-    for part in range(nPoints):
-        probMap = outputs[0,part,:,:]
-        probMap = cv2.resize(probMap, (456, 256)) # (456, 256)
-        keypoints = getKeypoints(probMap, threshold)
-        keypoints_with_id = []
-
-        for i in range(len(keypoints)):
-            keypoints_with_id.append(keypoints[i] + (keypoint_id,))
-            keypoints_list = np.vstack([keypoints_list, keypoints[i]])
-            keypoint_id += 1
-
-        detected_keypoints.append(keypoints_with_id)
-
-    valid_pairs, invalid_pairs = getValidPairs(outputs, 456, 256,detected_keypoints)
-    personwiseKeypoints = getPersonwiseKeypoints(valid_pairs, invalid_pairs,keypoints_list)
-    keypoints_limbs = [detected_keypoints,personwiseKeypoints,keypoints_list]
-
-    return keypoints_limbs
-
-def show_openpose(keypoints_limbs, frame, **kwargs):
-    frame = np.uint8(frame.copy())
-    if len(keypoints_limbs) == 3:
-        detected_keypoints = keypoints_limbs[0]
-        personwiseKeypoints = keypoints_limbs[1]
-        keypoints_list = keypoints_limbs[2]
-
-        for i in range(nPoints):
-            for j in range(len(detected_keypoints[i])):
-                cv2.circle(frame, detected_keypoints[i][j][0:2], 5, colors[i], -1, cv2.LINE_AA)
-        for i in range(17):
-            for n in range(len(personwiseKeypoints)):
-                index = personwiseKeypoints[n][np.array(POSE_PAIRS[i])]
-                if -1 in index:
-                    continue
-                B = np.int32(keypoints_list[index.astype(int), 0])
-                A = np.int32(keypoints_list[index.astype(int), 1])
-                cv2.line(frame, (B[0], A[0]), (B[1], A[1]), colors[i], 3, cv2.LINE_AA)
-    return frame
-
