--- gen2_helpers.py
@@ -1,17 +1,0 @@
-import cv2
-import numpy as np
-
-
-def cos_dist(a, b):
-    return np.dot(a, b) / (np.linalg.norm(a) * np.linalg.norm(b))
-
-
-def frame_norm(frame, bbox):
-    norm_vals = np.full(len(bbox), frame.shape[0])
-    norm_vals[::2] = frame.shape[1]
-    return (np.clip(np.array(bbox), 0, 1) * norm_vals).astype(int)
-
-def to_planar(arr: np.ndarray, shape: tuple = None) -> np.ndarray:
-    if shape is None:
-        return arr.transpose(2, 0, 1)
-    return cv2.resize(arr, shape).transpose(2, 0, 1)
