--- model_compiler/.gitignore
@@ -1,2 +1,0 @@
-downloads/
-models.txt
