--- model_compiler/downloader/requirements.in
@@ -1,2 +1,0 @@
-pyyaml
-requests
