+++ requirements.txt
@@ -4,3 +4,4 @@ requests==2.24.0
 argcomplete==1.12.1
 pyyaml==5.3.1
 depthai==2.1.0.0
+blobconverter==0.0.1
