--- resources/nn/age-gender-recognition-retail-0013/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Age & gender classification. Used in Audience Analytics.  
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_age_gender_recognition_retail_0013_description_age_gender_recognition_retail_0013.html
-task_type: object_attributes
-files:
-  - name: FP32/age-gender-recognition-retail-0013.xml
-    size: 30006
-    sha256: ca2c4826b9370038643d97a434bb9621adea4baf4279e5b9c3a83002456fd755
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP32/age-gender-recognition-retail-0013.xml
-  - name: FP32/age-gender-recognition-retail-0013.bin
-    size: 8552076
-    sha256: 4dab79cfedebd628f3327367a91c9736a32fbf9cc733cbbf1629d16910d4ace7
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP32/age-gender-recognition-retail-0013.bin
-  - name: FP16/age-gender-recognition-retail-0013.xml
-    size: 29996
-    sha256: a375a5d8617ff1913dfc1210f5bb07493d72293c0efa87c1abcdfe181ae8a2b0
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP16/age-gender-recognition-retail-0013.xml
-  - name: FP16/age-gender-recognition-retail-0013.bin
-    size: 4276038
-    sha256: 401101e0a01b3d68add39deae833bcf9f54238d37dd13e3fb1534aa8fbe4719d
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP16/age-gender-recognition-retail-0013.bin
-  - name: FP32-INT8/age-gender-recognition-retail-0013.xml
-    size: 64387
-    sha256: 930b5e2ef06d890a79e91f16ecbe54ce08abcb6923cfb56afa464750588d7a21
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP32-INT8/age-gender-recognition-retail-0013.xml
-  - name: FP32-INT8/age-gender-recognition-retail-0013.bin
-    size: 8565952
-    sha256: 0efcef2c78fa5208d85d2eb5fc69955c6d97a2ba2948c67a75b0dcb3b74da41e
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/age-gender-recognition-retail-0013/FP32-INT8/age-gender-recognition-retail-0013.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
