--- resources/nn/face-detection-adas-0001/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Face Detection (MobileNet with reduced channels + SSD with weights sharing)
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_face_detection_adas_0001_description_face_detection_adas_0001.html
-task_type: detection
-files:
-  - name: FP32/face-detection-adas-0001.xml
-    size: 232998
-    sha256: 62b4bf7dead77e16a47428b541aa4f3c506cdf3c7e31a317aa75771dd907557c
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP32/face-detection-adas-0001.xml
-  - name: FP32/face-detection-adas-0001.bin
-    size: 4212072
-    sha256: 85a9334e031289692884e2aefbcb4ca401b003a3f25ff4dd0e669ba32f98cc0b
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP32/face-detection-adas-0001.bin
-  - name: FP16/face-detection-adas-0001.xml
-    size: 232967
-    sha256: c0b9d34eba1fe6c76755fada4dc65634068eddf343a219b501ac71439348eac1
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP16/face-detection-adas-0001.xml
-  - name: FP16/face-detection-adas-0001.bin
-    size: 2106088
-    sha256: df0f5799d801c6afb355d1c4771693c782efbb58d2eb2238982eac8fe84bc821
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP16/face-detection-adas-0001.bin
-  - name: FP32-INT8/face-detection-adas-0001.xml
-    size: 503962
-    sha256: 8959f287e3caf098948b2faee40d67e019c4b794948476fcdc8a28e58cb79125
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP32-INT8/face-detection-adas-0001.xml
-  - name: FP32-INT8/face-detection-adas-0001.bin
-    size: 4291532
-    sha256: d3ffb0da33361253d4027def5acddf104a6919c920582f084974a4f9d8529ffd
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/face-detection-adas-0001/FP32-INT8/face-detection-adas-0001.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
