--- resources/nn/facial-landmarks-35-adas-0002/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Custom-architecture convolutional neural network for 35 facial landmarks estimation.
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_facial_landmarks_35_adas_0002_description_facial_landmarks_35_adas_0002.html
-task_type: object_attributes
-files:
-  - name: FP32/facial-landmarks-35-adas-0002.xml
-    size: 237139
-    sha256: 4cb90657a60311184f3505ef973563180729bb7e8d5a2f6a42f9109c1af7bfdb
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP32/facial-landmarks-35-adas-0002.xml
-  - name: FP32/facial-landmarks-35-adas-0002.bin
-    size: 18381152
-    sha256: bd41c25201c2ea688a75549c11b9e10aa98e51dcc5a6dd783e84cdcf55a1011e
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP32/facial-landmarks-35-adas-0002.bin
-  - name: FP16/facial-landmarks-35-adas-0002.xml
-    size: 237002
-    sha256: 702d476efcab1e89519ff14b76812c61cd5123f86f5fc343512fa416a9a94e27
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP16/facial-landmarks-35-adas-0002.xml
-  - name: FP16/facial-landmarks-35-adas-0002.bin
-    size: 9190584
-    sha256: 88a2064849da0e6a31557f489ca4d652aa99e5a9676da0c5c857ee5d8cc26fe4
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP16/facial-landmarks-35-adas-0002.bin
-  - name: FP32-INT8/facial-landmarks-35-adas-0002.xml
-    size: 588492
-    sha256: d41ba7372ca2db93bd269db32567022109d3e2a9b7aed3da0ef1e8e021f95be2
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP32-INT8/facial-landmarks-35-adas-0002.xml
-  - name: FP32-INT8/facial-landmarks-35-adas-0002.bin
-    size: 18441156
-    sha256: f81ebf7c25f30138c2fa44530f51f516783ea790c5b6cacbaa5b31d95c73cbb7
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/facial-landmarks-35-adas-0002/FP32-INT8/facial-landmarks-35-adas-0002.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
