--- resources/nn/human-pose-estimation-0001/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  2D human pose estimation with tuned mobilenet backbone (based on OpenPose).
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_human_pose_estimation_0001_description_human_pose_estimation_0001.html
-task_type: human_pose_estimation
-files:
-  - name: FP32/human-pose-estimation-0001.xml
-    size: 143782
-    sha256: 4d2cc5d3a8c395affa74ae640a63ba9bae9bd5871527d65f62a87cc5b40831e9
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP32/human-pose-estimation-0001.xml
-  - name: FP32/human-pose-estimation-0001.bin
-    size: 16394712
-    sha256: 1cd32cb5f9f4633b8ee1451144f749ddd75038c4c5fe862e704e489908b79a78
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP32/human-pose-estimation-0001.bin
-  - name: FP16/human-pose-estimation-0001.xml
-    size: 143706
-    sha256: 75bda095e6fa1b1c118e085c4e5f4be9d9d9db512578fd6ec57b3113d5b42dae
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP16/human-pose-estimation-0001.xml
-  - name: FP16/human-pose-estimation-0001.bin
-    size: 8197356
-    sha256: c1c828d5d1ea2b03035692a8600f06d787fe1a20e79dac159c4293ed7063a382
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP16/human-pose-estimation-0001.bin
-  - name: FP32-INT8/human-pose-estimation-0001.xml
-    size: 408100
-    sha256: 6e22b77c4e5f2c39bb14603c3bfa2d902210a3fdb3fe89443eab5fb149c3a470
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP32-INT8/human-pose-estimation-0001.xml
-  - name: FP32-INT8/human-pose-estimation-0001.bin
-    size: 16513820
-    sha256: 482635ee949716049bd7dd5a69ab66a6d5f29a5cc86598c0369035be2d1fa274
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/human-pose-estimation-0001/FP32-INT8/human-pose-estimation-0001.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
