--- resources/nn/mobilenet-ssd/model.yml
@@ -1,43 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  The `mobilenet-ssd` model is a Single-Shot multibox Detection (SSD) network
-  intended to perform object detection. This model is implemented using the Caffe\*
-  framework. For details about this model, check out the repository <https://github.com/chuanqi305/MobileNet-SSD>.
-
-  The model input is a blob that consists of a single image of 1x3x300x300 in
-  BGR order, also like the `densenet-121` model. The BGR mean values need to be
-  subtracted as follows: [127.5, 127.5, 127.5] before passing the image blob into
-  the network. In addition, values must be divided by 0.007843.
-
-  The model output is a typical vector containing the tracked object data, as
-  previously described.
-documentation: https://github.com/openvinotoolkit/open_model_zoo/blob/efd238d02035f8a5417b7b1e25cd4c997d44351f/models/public/mobilenet-ssd/mobilenet-ssd.md
-task_type: detection
-files:
-  - name: FP16/mobilenet-ssd.xml
-    size: 177219
-    sha256: ec3a3931faf1a4a5d70aa12576cc2a5f1b5b0d0be2517cc8f9c42f616fa10b2f
-    source:
-      $type: google_drive
-      id: 11-PX4EDxAnhymbuvnyb91ptvZAW3oPOn
-  - name: FP16/mobilenet-ssd.bin
-    size: 11566946
-    sha256: db075a98c8d3e4636bb4206048b3d20f164a5a3730a5aa6b6b0cdbbfd2607fab
-    source:
-      $type: google_drive
-      id: 1pdC4eNWxyfewCJ7T0i9SXLHKt39gBDZV
-framework: dldt
-license: https://raw.githubusercontent.com/chuanqi305/MobileNet-SSD/master/LICENSE
