--- resources/nn/openpose2
@@ -1,1 +1,0 @@
-mobileNetV2-PoseEstimation/
