--- resources/nn/person-detection-retail-0013/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Pedestrian detection (RMNet with lrelu + SSD)
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_person_detection_retail_0013_description_person_detection_retail_0013.html
-task_type: detection
-files:
-  - name: FP32/person-detection-retail-0013.xml
-    size: 354260
-    sha256: d83a9b8a82a9e6fb52c73f63c4ab6e71d57592c8b2068a85f14548df7cef8eae
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP32/person-detection-retail-0013.xml
-  - name: FP32/person-detection-retail-0013.bin
-    size: 2891364
-    sha256: 6f09cb7061328942f9d5e9fc81631a4234be66a26daa50cd672d4077ee82ad44
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP32/person-detection-retail-0013.bin
-  - name: FP16/person-detection-retail-0013.xml
-    size: 354107
-    sha256: 603502ad82af0ee6b0ee9948257aece2599ac2b153df1691b3fe5b0c33fe3a88
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP16/person-detection-retail-0013.xml
-  - name: FP16/person-detection-retail-0013.bin
-    size: 1445736
-    sha256: 56eeccbeb3f27144046edacb95b459d60f3930f54051ef5b9e5253923c07b672
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP16/person-detection-retail-0013.bin
-  - name: FP32-INT8/person-detection-retail-0013.xml
-    size: 931279
-    sha256: 8f7e71c2db19f1f641346ff9cd99553e650b42c00a96c9bbfaf2bcbbd6daa49d
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP32-INT8/person-detection-retail-0013.xml
-  - name: FP32-INT8/person-detection-retail-0013.bin
-    size: 2969676
-    sha256: 523b79294453845934e068f4a906b9bf88b4c70ab705f2137e089439f4555b92
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-detection-retail-0013/FP32-INT8/person-detection-retail-0013.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
