--- resources/nn/person-vehicle-bike-detection-crossroad-1016/model.yml
@@ -1,46 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Multiclass (person -  vehicle -  non-vehicle) detector based on SSD detection architecture
-  -  MobileNetV2 backbone)
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_person_vehicle_bike_detection_crossroad_1016_description_person_vehicle_bike_detection_crossroad_1016.html
-task_type: detection
-files:
-  - name: FP32/person-vehicle-bike-detection-crossroad-1016.xml
-    size: 194333
-    sha256: 040bded9370548d6ca02286b764334291c4ce6edb3ca6facd1758490a6433230
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP32/person-vehicle-bike-detection-crossroad-1016.xml
-  - name: FP32/person-vehicle-bike-detection-crossroad-1016.bin
-    size: 11548004
-    sha256: 61c9deaca903bb107aecfc31faeb402c894e0d43ab3817db8d9d05d43dfa68cf
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP32/person-vehicle-bike-detection-crossroad-1016.bin
-  - name: FP16/person-vehicle-bike-detection-crossroad-1016.xml
-    size: 194219
-    sha256: 3392a26b381ebbfa837cafb976a828fd8519ff58be08aac72cb54f4f40b58184
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP16/person-vehicle-bike-detection-crossroad-1016.xml
-  - name: FP16/person-vehicle-bike-detection-crossroad-1016.bin
-    size: 5774054
-    sha256: a0ba3bc9ca17330c731b395153651c22a666050cd06413c01059070681dc79a7
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP16/person-vehicle-bike-detection-crossroad-1016.bin
-  - name: FP32-INT8/person-vehicle-bike-detection-crossroad-1016.xml
-    size: 510391
-    sha256: 82d95cf77a1a5a4c2d0d8a6c364229700f293ccd8df883a081c2dc33f4025462
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP32-INT8/person-vehicle-bike-detection-crossroad-1016.xml
-  - name: FP32-INT8/person-vehicle-bike-detection-crossroad-1016.bin
-    size: 11777480
-    sha256: 0abdc201405b2f910f3ddd49307afd89ee43ab80009f9827b4f16e7aa7274157
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/person-vehicle-bike-detection-crossroad-1016/FP32-INT8/person-vehicle-bike-detection-crossroad-1016.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
