--- resources/nn/vehicle-license-plate-detection-barrier-0106/model.yml
@@ -1,45 +1,0 @@
-# Copyright (c) 2019 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#      http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-description: >-
-  Multiclass (vehicle -  license plates) detector based on MobileNetV2+SSD
-documentation: https://docs.openvinotoolkit.org/2020.1/_models_intel_vehicle_license_plate_detection_barrier_0106_description_vehicle_license_plate_detection_barrier_0106.html
-task_type: detection
-files:
-  - name: FP32/vehicle-license-plate-detection-barrier-0106.xml
-    size: 204476
-    sha256: 0c4abcf31945d04fc33459625988176ade8e0986a8e2b4499c1ec511177217a4
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP32/vehicle-license-plate-detection-barrier-0106.xml
-  - name: FP32/vehicle-license-plate-detection-barrier-0106.bin
-    size: 2573380
-    sha256: 2a6f1cd0eb8731ef059a91299bce0be026d2cebb1ae42cbee37b6aef7da2764b
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP32/vehicle-license-plate-detection-barrier-0106.bin
-  - name: FP16/vehicle-license-plate-detection-barrier-0106.xml
-    size: 204408
-    sha256: 6fc62d36f52ccfa34eced7c8e3c260e9870a8b5c42fe413a76a1321f10a637dc
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP16/vehicle-license-plate-detection-barrier-0106.xml
-  - name: FP16/vehicle-license-plate-detection-barrier-0106.bin
-    size: 1286758
-    sha256: 57652e098f93ee712eb412267612cb7bf2713475f96ec1e18eca10e4bef69785
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP16/vehicle-license-plate-detection-barrier-0106.bin
-  - name: FP32-INT8/vehicle-license-plate-detection-barrier-0106.xml
-    size: 519016
-    sha256: 0037685af2cc25f569c7db066ab6e9673a0faae204d8f2f19a003b28bd21bb28
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP32-INT8/vehicle-license-plate-detection-barrier-0106.xml
-  - name: FP32-INT8/vehicle-license-plate-detection-barrier-0106.bin
-    size: 2642612
-    sha256: 2f81d5803d76e90f9000c17083536ece01e7127c407e5bec7a6caef07ac01082
-    source: https://download.01.org/opencv/2020/openvinotoolkit/2020.1/open_model_zoo/models_bin/1/vehicle-license-plate-detection-barrier-0106/FP32-INT8/vehicle-license-plate-detection-barrier-0106.bin
-framework: dldt
-license: https://raw.githubusercontent.com/opencv/open_model_zoo/master/LICENSE
